lambda_functions/authorizer/requirements.txt
boto3
sqlalchemy
pyjwt